dir hot fix
