+++ src/main/java/de/uni_mannheim/informatik/dws/wdi/WDI_Project_T2/IdentityResolution/IR_using_linear_combination.java
@@ -1,7 +1,7 @@
 package de.uni_mannheim.informatik.dws.wdi.WDI_Project_T2.IdentityResolution;
 
 import de.uni_mannheim.informatik.dws.wdi.WDI_Project_T2.blocking.CarBlockingKeyByManufacturerGenerator;
-import de.uni_mannheim.informatik.dws.wdi.WDI_Project_T2.Comparators.CarManufacturerComparator_LowerCase;
+import de.uni_mannheim.informatik.dws.wdi.WDI_Project_T2.comparator.CarManufacturerComparator_LowerCase;
 import de.uni_mannheim.informatik.dws.wdi.WDI_Project_T2.model.Car;
 import de.uni_mannheim.informatik.dws.wdi.WDI_Project_T2.model.CarXMLReader;
 import de.uni_mannheim.informatik.dws.winter.matching.MatchingEngine;
