+++ src/main/java/de/uni_mannheim/informatik/dws/wdi/WDI_Project_T2/comparator/Car_Model_TokenizingJaccard_Comparator.java
@@ -1,4 +1,4 @@
-package de.uni_mannheim.informatik.dws.wdi.WDI_Project_T2.Comparators;
+package de.uni_mannheim.informatik.dws.wdi.WDI_Project_T2.comparator;
 import de.uni_mannheim.informatik.dws.wdi.WDI_Project_T2.Utils;
 import de.uni_mannheim.informatik.dws.winter.matching.rules.Comparator;
 import de.uni_mannheim.informatik.dws.wdi.WDI_Project_T2.model.Car;
